feature/refactor-template-for-testing-and-reports
Note that this is 'using method Squash and Merge'. This is to keep a clean master branch history. I won't be deleting the branch though (as a work record).

Also note the naming convention for features:
feature/and-a-name-seperated-with-hyphens

See the following article: https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/about-pull-request-merges